chore(nft-profile): Removal and migration of profile code
