+++ src/views/Nft/market/Profile/components/Achievements/index.tsx
@@ -3,6 +3,7 @@ import { Card, CardBody, Heading, PrizeIcon } from '@pancakeswap/uikit'
 import { useTranslation } from 'contexts/Localization'
 import StatBox from './StatBox'
 import AchievementsList from './AchievementsList'
+import ClaimPointsCallout from './ClaimPointsCallout '
 
 const Achievements = ({ points = 0 }) => {
   const { t } = useTranslation()
@@ -14,6 +15,7 @@ const Achievements = ({ points = 0 }) => {
         <Heading as="h4" scale="md" mb="16px">
           {t('Achievements')}
         </Heading>
+        <ClaimPointsCallout />
         <AchievementsList />
       </CardBody>
     </Card>
