+++ src/views/Nft/market/Profile/components/EditProfileModal/PauseProfileView.tsx
@@ -1,7 +1,7 @@
 import React, { useState } from 'react'
 import { AutoRenewIcon, Button, Checkbox, Flex, InjectedModalProps, Text } from '@pancakeswap/uikit'
 import { useTranslation } from 'contexts/Localization'
-import useGetProfileCosts from 'views/Profile/hooks/useGetProfileCosts'
+import useGetProfileCosts from 'views/Nft/market/Profile/hooks/useGetProfileCosts'
 import { useAppDispatch } from 'state'
 import { useProfile } from 'state/profile/hooks'
 import { fetchProfile } from 'state/profile'
