+++ src/views/Nft/market/Profile/components/EditProfileModal/TeamAndProfileAvatar.tsx
