--- src/views/Profile/ProfileCreation/contexts/hook.ts
@@ -1,8 +1,0 @@
-import { useContext } from 'react'
-import { ProfileCreationContext } from './ProfileCreationProvider'
-
-const useProfileCreation = () => {
-  return useContext(ProfileCreationContext)
-}
-
-export default useProfileCreation
