--- src/views/Profile/components/ConfirmProfileCreationModal.tsx
@@ -1,86 +1,0 @@
-import React from 'react'
-import { Modal, Flex, Text } from '@pancakeswap/uikit'
-import { ethers } from 'ethers'
-import { formatUnits } from '@ethersproject/units'
-import { useAppDispatch } from 'state'
-import { useTranslation } from 'contexts/Localization'
-import { useCake, useProfile } from 'hooks/useContract'
-import useApproveConfirmTransaction from 'hooks/useApproveConfirmTransaction'
-import { fetchProfile } from 'state/profile'
-import useToast from 'hooks/useToast'
-import { useCallWithGasPrice } from 'hooks/useCallWithGasPrice'
-import { ToastDescriptionWithTx } from 'components/Toast'
-import { REGISTER_COST } from '../ProfileCreation/config'
-import ApproveConfirmButtons from './ApproveConfirmButtons'
-import { State } from '../ProfileCreation/contexts/types'
-
-interface Props {
-  userName: string
-  selectedNft: State['selectedNft']
-  account: string
-  teamId: number
-  minimumCakeRequired: ethers.BigNumber
-  allowance: ethers.BigNumber
-  onDismiss?: () => void
-}
-
-const ConfirmProfileCreationModal: React.FC<Props> = ({
-  account,
-  teamId,
-  selectedNft,
-  minimumCakeRequired,
-  allowance,
-  onDismiss,
-}) => {
-  const { t } = useTranslation()
-  const profileContract = useProfile()
-  const dispatch = useAppDispatch()
-  const { toastSuccess } = useToast()
-  const cakeContract = useCake()
-  const { callWithGasPrice } = useCallWithGasPrice()
-
-  const { isApproving, isApproved, isConfirmed, isConfirming, handleApprove, handleConfirm } =
-    useApproveConfirmTransaction({
-      onRequiresApproval: async () => {
-        try {
-          const response = await cakeContract.allowance(account, profileContract.address)
-          return response.gte(minimumCakeRequired)
-        } catch (error) {
-          return false
-        }
-      },
-      onApprove: () => {
-        return callWithGasPrice(cakeContract, 'approve', [profileContract.address, allowance.toJSON()])
-      },
-      onConfirm: () => {
-        return callWithGasPrice(profileContract, 'createProfile', [teamId, selectedNft.nftAddress, selectedNft.tokenId])
-      },
-      onSuccess: async ({ receipt }) => {
-        await dispatch(fetchProfile(account))
-        onDismiss()
-        toastSuccess(t('Profile created!'), <ToastDescriptionWithTx txHash={receipt.transactionHash} />)
-      },
-    })
-
-  return (
-    <Modal title={t('Complete Profile')} onDismiss={onDismiss}>
-      <Text color="textSubtle" mb="8px">
-        {t('Submitting NFT to contract and confirming User Name and Team.')}
-      </Text>
-      <Flex justifyContent="space-between" mb="16px">
-        <Text>{t('Cost')}</Text>
-        <Text>{t('%num% CAKE', { num: formatUnits(REGISTER_COST) })}</Text>
-      </Flex>
-      <ApproveConfirmButtons
-        isApproveDisabled={isConfirmed || isConfirming || isApproved}
-        isApproving={isApproving}
-        isConfirmDisabled={!isApproved || isConfirmed}
-        isConfirming={isConfirming}
-        onApprove={handleApprove}
-        onConfirm={handleConfirm}
-      />
-    </Modal>
-  )
-}
-
-export default ConfirmProfileCreationModal
