+++ src/views/Profile/components/EditProfileAvatar.tsx
@@ -1,7 +1,7 @@
 import React from 'react'
 import styled from 'styled-components'
 import { useModal } from '@pancakeswap/uikit'
-import EditProfileModal from './EditProfileModal'
+import EditProfileModal from 'views/Nft/market/Profile/components/EditProfileModal'
 import ProfileAvatar, { ProfileAvatarProps } from './ProfileAvatar'
 
 const EditOverlay = styled.div`
