--- src/views/Profile/components/HeaderWrapper.tsx
@@ -1,9 +1,0 @@
-import styled from 'styled-components'
-
-const HeaderWrapper = styled.div`
-  border-bottom: 2px solid ${({ theme }) => theme.colors.textSubtle};
-  margin-bottom: 24px;
-  padding-bottom: 24px;
-`
-
-export default HeaderWrapper
